fix(mobile): fix 235 analyzer errors across features batch 3
- 28 files modified across advisor, market, notifications, reports,
  irrigation, crops, auth, home, crm, spray, equipment, tasks
- Add explicit casts for strict-casts compliance
- Fix undefined getters/methods with proper imports and computed getters
- Fix ambiguous imports with hide directives
- Fix FarmerAnalytics missing computed properties
- Fix AppLogger.w() invalid error parameter

+++ apps/mobile/lib/core/analytics/user_properties.dart
@@ -164,7 +164,7 @@ class AnalyticsUserProperties {
     if (Platform.isAndroid) {
       final androidInfo = await deviceInfo.androidInfo;
       osVersion = 'Android ${androidInfo.version.release}';
-      deviceCategory = _getDeviceCategory(androidInfo.data['displayMetrics']?['widthPx']?.toDouble() ?? 1080);
+      deviceCategory = _getDeviceCategory((androidInfo.data['displayMetrics']?['widthPx']?.toDouble() ?? 1080) as double);
     } else if (Platform.isIOS) {
       final iosInfo = await deviceInfo.iosInfo;
       osVersion = 'iOS ${iosInfo.systemVersion}';
